Dynamic test factory for TestNG so that we can keep the same idiom for js tests
Test factory and corresponding test instance allows us to keep the simple "add a js
file and its corresponding out file" to add js tests.  Tested in the engine package
and will go add elsewhere (I think there must be others).

Tested that it fails correctly when an out is modified.

All 63 tests pass with this.  (There are 64 in home-build framework)