Get SiteSystemState through Scope.getThreadLocal.
Previously we got the threadlocal Python state and use that to map back to the
SiteSystemState that wraps it. Turns out, though, that this fails if, from JS,
you call a Python file, and then call a Python-defined function which does an
import. The PySystemState got restored (to be the vanilla default state) and
the import can't find the SiteSystemState it wants to be part of, so everything
goes sour.

Instead this way we get the currently running Scope and go straight from there
to the current SiteSystemState. The disadvantage is that going through the Scope
feels a bit clumsy and dirty, but the alternatives are probably just as bad and
more work.

Yet to do: Test, polish, etc. This is a hotfix to deal with shopwiki
breakage. My tests suggest that this shouldn't break the build, but
I've been wrong before.