Missing package declaration which doesn't seem to normally break build.
Except that when you're running Eclipse, it does break the build, because
Eclipse is smart enough to discover that the package declaration doesn't match
the directory name. And since the stupid language mandates that you repeat this
information, it stops processing immediately and loses its lunch all over your
shoes.