If there are more than 28 parameters, don't append a Object[0] to the end
Fixes (most of) Mozilla test ecma/Array/15.4.2.1-3.js.  Code like:

new Array(0,1,2,3,4,5,6,7,8,9,10,...,1024)

messes up because the converter adds ", Object[0]" to the generated Java and there's no JSFunctionCall for 1024 args, so the args are converted to an array with one junk element at the end.

I added a parameter counter, and if it goes above 28, the converter won't add the final Object[0].