Work around broken Jython madness.
Jython treats Thread interrupts as a signal to "restart". This sucks
because we obviously have our own ideas what an interrupt means. But
it sucks even more than that because the SystemRestart class doesn't
get generated -- at all, as far as I can tell. This is most likely
because the SystemRestart code is extremely experimental, so in the
meantime they generate PyExceptions that NPE when you call
printStackTrace().

This code is an extremely ham-handed way of saying "Hey Jython, create
your stupid class already".